Fix build on Windows. Now requires boost, so things are more complicated. Added a build script that works on my PC as an example.

+++ src/plugin_custom_dataref.cpp
@@ -36,4 +36,4 @@ PLUGIN_API int XPluginStart(char * outName, char * outSig, char * outDesc) {
 PLUGIN_API void	XPluginStop(void) { }
 PLUGIN_API void XPluginDisable(void) { }
 PLUGIN_API int XPluginEnable(void) { return 1; }
-PLUGIN_API void XPluginReceiveMessage(XPLMPluginID inFromWho, long inMessage, void * inParam) { }
+PLUGIN_API void XPluginReceiveMessage(XPLMPluginID inFromWho, intptr_t inMessage, void * inParam) { }
